+++ src/lte/test/test-sl-in-covrg-2relay-1remote-regular.cc
@@ -404,7 +404,7 @@ SlInCovrg2Relay1RemoteRegularTestCase::DoRun()
 
     if (relayTraffic)
     {
-        for (uint16_t relUeIdx = 0; relUeIdx < relayUeNodes.GetN(); relUeIdx++)
+        for (uint32_t relUeIdx = 0; relUeIdx < relayUeNodes.GetN(); relUeIdx++)
         {
             uint16_t relUePort = echoPortBase + 100 + relUeIdx;
 
@@ -434,7 +434,7 @@ SlInCovrg2Relay1RemoteRegularTestCase::DoRun()
 
     if (remoteTraffic)
     {
-        for (uint16_t remUeIdx = 0; remUeIdx < remoteUeNodes.GetN(); remUeIdx++)
+        for (uint32_t remUeIdx = 0; remUeIdx < remoteUeNodes.GetN(); remUeIdx++)
         {
             uint16_t remUePort = echoPortBase + remUeIdx;
